requirements.txt
torch==1.12.1+cu116 
torchvision==0.13.1+cu116
torchaudio==0.12.1
numpy==1.23.5
pandas==1.5.1
open_clip_torch
transformers
huggingface_hub
-e git+https://github.com/KaiyangZhou/Dassl.pytorch.git#egg=dassl
